Hacky fix to issue when total score column is wide.
When the total score is longer than 5 digits, NCBI outputs scientific
notation. Unfortunately, this shifts the rows over in the tables,
which causes a mismatch when reading in the data.

This solution just hardcodes in a case for when this happens, which is
not an ideal solution. Ideally, we would detect the column spacing for
each table dynamically (and therefore be robust to any other
variations coming out of NCBI). We might eventually do this, but for
now this quick solution should allow forward progress.