Updates the extract function to dynamically discover the col width.
The previous fix did not work for all cases, meaning that there are
more configurations of tables than just the two that we
hardcoded. This is an issues for the quick fix, and it is not
practical to hardcode all variations (as there are likely many we have
not seen before).

Therefore, we have to discover the column widths for each and every table.